+++ src/main/java/de/samply/template/ConverterTemplate.java
@@ -51,6 +51,10 @@ public class ConverterTemplate {
     return excelFilename;
   }
 
+  public String replaceTokenAndGetExcelFilename(){
+    return ConverterTemplateUtils.replaceTokens(excelFilename);
+  }
+
   public void setExcelFilename(String excelFilename) {
     this.excelFilename = excelFilename;
   }