+++ src/main/java/de/samply/csv/ContainersToCsvConverter.java
@@ -7,6 +7,7 @@ import de.samply.converter.Format;
 import de.samply.template.ContainerTemplate;
 import de.samply.template.ConverterTemplate;
 import de.samply.teiler.TeilerConst;
+import de.samply.template.ConverterTemplateUtils;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -16,6 +17,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
@@ -23,21 +25,26 @@ import reactor.core.publisher.Flux;
 @Component
 public class ContainersToCsvConverter extends ConverterImpl<Containers, Path> {
 
+  private ConverterTemplateUtils converterTemplateUtils;
   private String writeDirectory;
 
-  public ContainersToCsvConverter(@Value(TeilerConst.WRITE_FILE_DIRECTORY_SV) String writeDirectory) {
+  public ContainersToCsvConverter(@Autowired ConverterTemplateUtils converterTemplateUtils,
+      @Value(TeilerConst.WRITE_FILE_DIRECTORY_SV) String writeDirectory) {
     this.writeDirectory = writeDirectory;
+    this.converterTemplateUtils = converterTemplateUtils;
   }
 
   @Override
   protected Flux<Path> convert(Containers containers, ConverterTemplate template) {
     Map<String, String> filenameMap = new HashMap<>();
     Flux<Path> pathFlux = Flux.empty();
-    writeContainersInCsv(containers, template, filenameMap).forEach(path -> pathFlux.concatWithValues(path));
+    writeContainersInCsv(containers, template, filenameMap).forEach(
+        path -> pathFlux.concatWithValues(path));
     return pathFlux;
   }
 
-  public List<Path> writeContainersInCsv(Containers containers, ConverterTemplate template, Map<String,String> filenameMap) {
+  public List<Path> writeContainersInCsv(Containers containers, ConverterTemplate template,
+      Map<String, String> filenameMap) {
     List<Path> pathList = new ArrayList<>();
     template.getContainerTemplates().forEach(containerTemplate ->
         pathList.add(writeContainersInCsv(containers.getContainers(containerTemplate),
@@ -47,16 +54,19 @@ public class ContainersToCsvConverter extends ConverterImpl<Containers, Path> {
   }
 
   private Path writeContainersInCsv(List<Container> containers,
-      ConverterTemplate converterTemplate, ContainerTemplate containerTemplate, Map<String,String> filenameMap) {
+      ConverterTemplate converterTemplate, ContainerTemplate containerTemplate,
+      Map<String, String> filenameMap) {
     try {
-      return writeContainersInCsvWithoutExceptionManagement(containers, converterTemplate, containerTemplate, filenameMap);
+      return writeContainersInCsvWithoutExceptionManagement(containers, converterTemplate,
+          containerTemplate, filenameMap);
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
   }
 
   private Path writeContainersInCsvWithoutExceptionManagement(List<Container> containers,
-      ConverterTemplate converterTemplate, ContainerTemplate containerTemplate, Map<String,String> filenameMap) throws IOException {
+      ConverterTemplate converterTemplate, ContainerTemplate containerTemplate,
+      Map<String, String> filenameMap) throws IOException {
 
     Path filePath = getFilePath(containerTemplate, filenameMap);
     boolean headersExists = headersExists(filePath);
@@ -68,10 +78,10 @@ public class ContainersToCsvConverter extends ConverterImpl<Containers, Path> {
 
   }
 
-  private Path getFilePath(ContainerTemplate containerTemplate, Map<String,String> filenameMap) {
+  private Path getFilePath(ContainerTemplate containerTemplate, Map<String, String> filenameMap) {
     String filename = filenameMap.get(containerTemplate.getCsvFilename());
-    if (filename == null){
-      filename = containerTemplate.replaceTokensAndGetCsvFilename();
+    if (filename == null) {
+      filename = converterTemplateUtils.replaceTokens(containerTemplate.getCsvFilename());
       filenameMap.put(containerTemplate.getCsvFilename(), filename);
     }
     return Paths.get(writeDirectory).resolve(filename);