+++ src/test/java/de/samply/template/ConverterTemplateTest.java
@@ -57,11 +57,12 @@ class ConverterTemplateTest {
     AttributeTemplate attributeTemplate = new AttributeTemplate(j, "col-" + extension,
         "col-" + extension,
         "path.to.attribute." + extension);
-    attributeTemplate.setParentFhirPath("path.to.parent.container-"+i);
+    attributeTemplate.setParentFhirPath("path.to.parent.container-" + i);
+    attributeTemplate.setMdr("urn:test:dataelement:" + i + ":" + j);
     return attributeTemplate;
   }
 
-  private List<String> generateFhirRevIncludes(){
+  private List<String> generateFhirRevIncludes() {
     List<String> fhirRevIncludes = new ArrayList<>();
     fhirRevIncludes.add("Observation:patient");
     fhirRevIncludes.add("Condition:patient");