+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Project Version
 - REST API: retrieve-query
 - Docker
+- Timestamp variable for templates
 
 ### Changed
 - Rename Conversion Template as Converter Template