+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Timestamp variable for templates
 - Environment variables for templates
 - Parent-fhir-path in attribute template
+- MDR URN in attribute template
 
 ### Changed
 - Rename Conversion Template as Converter Template